# swagger.yaml
openapi: 3.0.0
servers:
  - url: https://reservaksin-be.herokuapp.com
    description: ""
    variables: {}
info:
  version: v.1.0
  title: Reservaksin-API
  description: "The goal of the project is to help the government’s goal of Covid-19 vaccinations by facilitating users to\_**book vaccination sessions easily**\_and creating a proper system for health facilities."
  termsOfService: ""
  contact: {}
  license:
    name: ""
paths:
...
  /vaccine/admin/{id}:
    get:
      summary: GetByAdminID
      operationId: GetByAdminID
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: string
            example: nHBMa-88t78LWI9vrbCo3
      responses:
        "200":
          description: GetByAdminID [success]
          content:
            application/json; charset=UTF-8:
              schema:
                $ref: "#/components/schemas/200"
              example:
                meta:
                  status: 200
                  message: succes
                data:
                  - id: 4LtskWv7JmmjccAK28Cjk
                    nama_vaksin: Moderna
                    admin_id: luYbco9Im0yqrODt08KLE
                    stok: 1000
                    created_at: "0001-01-01T00:00:00Z"
                    updated_at: "2022-01-13T14:07:48+07:00"
                  - id: SE9why4D2PMmiU75T3OSc
                    nama_vaksin: Sinopharm
                    admin_id: luYbco9Im0yqrODt08KLE
                    stok: 900
                    created_at: "2022-01-13T13:59:59.81+07:00"
                    updated_at: "2022-01-13T13:59:59.81+07:00"
                  - id: UKNHAfxsMzIOGZUOoGQMF
                    nama_vaksin: AstraZeneca
                    admin_id: luYbco9Im0yqrODt08KLE
                    stok: 250
                    created_at: "0001-01-01T00:00:00Z"
                    updated_at: "2022-01-07T17:58:57.135+07:00"
                  - id: VWmoAFO5NshFGcOFjxIDH
                    nama_vaksin: Pfizer
                    admin_id: luYbco9Im0yqrODt08KLE
                    stok: 500
                    created_at: "2022-01-02T20:42:33.658+07:00"
                    updated_at: "2022-01-02T20:42:33.658+07:00"
                  - id: XtMM326PzZhcF-Z4aVJtC
                    nama_vaksin: Sinovac
                    admin_id: luYbco9Im0yqrODt08KLE
                    stok: 500
                    created_at: "2022-01-02T20:42:26.258+07:00"
                    updated_at: "2022-01-02T20:42:26.258+07:00"
      tags:
        - Vaccine
...
  /session/upcoming/admin/{id}:
    get:
      summary: Upcoming
      operationId: Upcoming
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: string
            example: bNQV1Uy9JtkoQIwERfyit
      responses:
        "200":
          description: Upcoming
          content:
            application/json; charset=UTF-8:
              schema:
                $ref: "#/components/schemas/200"
              example:
                meta:
                  status: 200
                  message: Data is Empty
                data: []
      tags:
        - Session
  /session/current/admin/{id}:
    get:
      summary: Current
      operationId: Current
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: string
            example: bNQV1Uy9JtkoQIwERfyit
      responses:
        "200":
          description: Current
          content:
            application/json; charset=UTF-8:
              schema:
                $ref: "#/components/schemas/200"
              example:
                meta:
                  status: 200
                  message: succes
                data: no current session
      tags:
        - Session
  /session/history/admin/{id}:
    get:
      summary: History
      operationId: History
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: string
            example: bNQV1Uy9JtkoQIwERfyit
      responses:
        "200":
          description: History
          content:
            application/json; charset=UTF-8:
              schema:
                $ref: "#/components/schemas/200"
              example:
                meta:
                  status: 200
                  message: succes
                data:
                  - id: yUxEwcOATwcq7vKBpZmz_
                    health_facilities_id: yUxEwcOATwcq7vKBpZmz_
                    health_facilities: RS. Siloam Hosiptal International
                    name_session: Vaksinasi Mantap
                    capacity: 500
                    vaccine_id: XtMM326PzZhcF-Z4aVJtC
                    vaccine: Sinovac
                    date: "2022-01-07T00:00:00+07:00"
                    tahap: "1"
                    start_session: "2022-01-07T08:00:00+07:00"
                    end_session: "2022-01-07T16:00:00+07:00"
                    created_at: "2022-01-05T19:04:26.321+07:00"
                    updated_at: "2022-01-05T19:05:07.518+07:00"
                  - id: vOPrWPXqRXkeoaQzmYRcx
                    health_facilities_id: vOPrWPXqRXkeoaQzmYRcx
                    health_facilities: RS. Mentari Pagi
                    name_session: Vaksin Gratis Warga
                    capacity: 500
                    vaccine_id: 4LtskWv7JmmjccAK28Cjk
                    vaccine: Moderna
                    date: "2022-01-07T00:00:00+07:00"
                    tahap: "1"
                    start_session: "2022-01-07T08:00:00+07:00"
                    end_session: "2022-01-07T16:00:00+07:00"
                    created_at: "2022-01-05T17:03:59.235+07:00"
                    updated_at: "2022-01-05T17:08:56.3+07:00"
      tags:
        - Session
  /session/admin/{id}:
    get:
      summary: FetchAllByAdminID
      operationId: FetchAllByAdminID
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: string
            example: AMjDWJ6YsewbWR36-oz-f
      responses:
        "200":
          description: FetchAllByAdminID
          content:
            application/json; charset=UTF-8:
              schema:
                $ref: "#/components/schemas/200"
              example:
                meta:
                  status: 200
                  message: succes
                data:
                  - id: yUxEwcOATwcq7vKBpZmz_
                    health_facilities_id: yUxEwcOATwcq7vKBpZmz_
                    health_facilities: RS. Siloam Hosiptal International
                    name_session: Vaksinasi Mantap
                    capacity: 500
                    vaccine_id: XtMM326PzZhcF-Z4aVJtC
                    vaccine: Sinovac
                    date: "2022-01-07T00:00:00+07:00"
                    tahap: "1"
                    start_session: "2022-01-07T08:00:00+07:00"
                    end_session: "2022-01-07T16:00:00+07:00"
                    created_at: "2022-01-05T19:04:26.321+07:00"
                    updated_at: "2022-01-05T19:05:07.518+07:00"
                  - id: vOPrWPXqRXkeoaQzmYRcx
                    health_facilities_id: vOPrWPXqRXkeoaQzmYRcx
                    health_facilities: RS. Mentari Pagi
                    name_session: Vaksin Gratis Warga
                    capacity: 500
                    vaccine_id: 4LtskWv7JmmjccAK28Cjk
                    vaccine: Moderna
                    date: "2022-01-07T00:00:00+07:00"
                    tahap: "1"
                    start_session: "2022-01-07T08:00:00+07:00"
                    end_session: "2022-01-07T16:00:00+07:00"
                    created_at: "2022-01-05T17:03:59.235+07:00"
                    updated_at: "2022-01-05T17:08:56.3+07:00"
      tags:
        - Session
...
components:
  schemas:
    "200":
      type: object
      properties:
        meta:
          type: object
          properties:
            status:
              type: integer
              description: ""
              default: 200
            message:
              type: string
              description: ""
              default: succes
        data:
          type: object
          properties:
            id:
              type: string
              description: ""
              default: luYbco9Im0yqrODt08KLE
            role:
              type: string
              description: ""
              default: admin
            username:
              type: string
              description: ""
              default: kurniawan12
            password:
              type: string
              description: ""
              default: $2a$04$FhY9ZzerGjL0Q4tlqpGdNuEcvh7mrgt1iLGp9JHO5MwGLtO6LtS3m
            created_at:
              type: string
              description: ""
              default: "2022-01-02T20:43:38.917+07:00"
            updated_at:
              type: string
              description: ""
              default: "2022-01-02T20:43:38.917+07:00"
...
security: []
...